+++ src/event.c
@@ -78,6 +78,8 @@ int event_record(int ev_type, int ev_code, void *ev_value)
 		case EV_UPGRADE_FAILED:
 		case EV_COMPLETE:
 		case EV_LOW_POWER:
+		case EV_USERDATA_ERROR_SKIP:
+		case EV_USERDATA_ERROR_MANDATORY:
 			len += snprintf(line_buffer + len, BUFSIZ - len, "%#x\n", ev_type);
 			//snprintf(line_buffer, BUFSIZ, "%#x,,,\n", ev_type);
 			break;